+++ src/main/java/com/quasiris/qsf/commons/text/normalizer/NormalizerConfig.java
@@ -2,7 +2,10 @@ package com.quasiris.qsf.commons.text.normalizer;
 
 /**
  * App Text Normalizer config
+ *
+ * @deprecated  As of release 0.4, replaced by {@link com.quasiris.qsf.commons.text.transform.TextTransformer}
  */
+@Deprecated
 public class NormalizerConfig {
     /**
      * Profile id